Update README.md
Final project for ITF Manual Testing Course
The scope of the final project for ITF Manual Testing Course is to use all gained knowledge through the course and apply them in practice, using a live application.

Application under test:

API Documentation:

The final project will be split into 2 sections: Testing section and SQL section.

Tools used:

Functional specifications
-> enter here the functional specifications created in JIRA

1 Testing section
1.1 Test Planning
The Test Plan is designed to describe all details of testing for the X module from the OrangeHRM application.

The plan identifies the items to be tested, the features to be tested, the types of testing to be performed, the personnel responsible for testing, the resources and schedule required to complete testing, and the risks associated with the plan

1.1.1 Roles assigned to the project and persons allocated
1.1.2 Entry criteria defined
1.1.3 Exit criteria defined
1.1.4 Test scope
Tests in scope:
Tests not in scope:
1.1.5 Risks detected
Project risks:
Product risks:
1.1.6 Evaluating entry criteria
The entry criteria defined in the Test Planning phase have been achieved and the test process can continue.

1.2 Test Monitoring and Control
It will be done by generating periodic reports that reflect the current status of the test.

1.3 Test Analysis
The testing process will be executed based on the above requirements for the Dependents module. The following test conditions were found:

Enter test conditions here
1.4 Test Design
Functional test cases were created in Zephyr Squad. Based on the analysis of the specifications, the test design techniques used for generating test cases are:

Test cases: -> enter here test cases or at least the titles

The test cases with steps can be viewed here: test_cases.pdf

1.5 Test Implementation
The following elements are needed to be ready before the test execution phase begins:

enter here what needs to be ready for the test execution to begin
1.6 Test Execution
Test cases are executed on the created test Cycle summary: cycle_summary_execution.pdf
Bugs have been created based on the failed tests. The complete bug reports can be found here: created_bugs.pdf
enter here bug titles
1.7 Test Completion
Exit criteria was evaluated and passed
The traceability matrix was generated and can be found here: Traceability_matrix.csv
Test execution chart was generated, the final report shows.... -> describe the final report
-> enter here test execution report/chart

2 SQL section